backend/requirements.txt
fastapi
uvicorn

# LLM
langchain==1.2.0
langchain-core
langchain-community
langchain-huggingface
sentence-transformers
langchain-openai
openai

# DB
psycopg2-binary
sqlalchemy
faiss-cpu
numpy
sentence-transformers
transformers
torch

# opensearch
opensearch-py>=2.4.0

# Config
pydantic
pydantic-settings

# Utils
python-dotenv
